requirements.txt
flask
requests
flask-jwt-extended
python-dotenv
Flask-Selfdoc
gunicorn
mysql-connector-python
